# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart RO purifier API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SensorReadingPumpState =
  (typeof SensorReadingPumpState)[keyof typeof SensorReadingPumpState];

export const SensorReadingPumpState = {
  ON: "ON",
  OFF: "OFF",
} as const;

export type SensorReadingAlert =
  (typeof SensorReadingAlert)[keyof typeof SensorReadingAlert];

export const SensorReadingAlert = {
  NORMAL: "NORMAL",
  HIGH_TEMPERATURE: "HIGH_TEMPERATURE",
} as const;

export interface SensorReading {
  id: string;
  timestamp: string;
  temperature: number;
  tds: number;
  waterLevel: number;
  pumpState: SensorReadingPumpState;
  threshold: number;
  alert: SensorReadingAlert;
  manualMode: boolean;
}

export type AlertEventSeverity =
  (typeof AlertEventSeverity)[keyof typeof AlertEventSeverity];

export const AlertEventSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface AlertEvent {
  id: string;
  timestamp: string;
  severity: AlertEventSeverity;
  message: string;
  reading?: SensorReading;
}

export interface MqttConnectionStatus {
  configured: boolean;
  connected: boolean;
  brokerUrl?: string;
}

export interface DatabaseConnectionStatus {
  configured: boolean;
  connected: boolean;
  provider: string;
}

export type HardwarePinMapTemperatureSensor = {
  sensor: string;
  dataPin: string;
};

export type HardwarePinMapWaterLevelSensor = {
  analogPin: string;
};

export type HardwarePinMapTdsSensor = {
  analogPin: string;
};

export type HardwarePinMapRelay = {
  inputPin: string;
};

export type HardwarePinMapOledDisplay = {
  sclPin: string;
  sdaPin: string;
};

export interface HardwarePinMap {
  temperatureSensor: HardwarePinMapTemperatureSensor;
  waterLevelSensor: HardwarePinMapWaterLevelSensor;
  tdsSensor: HardwarePinMapTdsSensor;
  relay: HardwarePinMapRelay;
  oledDisplay: HardwarePinMapOledDisplay;
}

export type SystemStatusMode =
  (typeof SystemStatusMode)[keyof typeof SystemStatusMode];

export const SystemStatusMode = {
  AUTOMATIC: "AUTOMATIC",
  MANUAL_READY: "MANUAL_READY",
} as const;

export interface SystemStatus {
  mode: SystemStatusMode;
  threshold: number;
  pumpRule: string;
  latest: SensorReading;
  mqtt: MqttConnectionStatus;
  database: DatabaseConnectionStatus;
  topics: string[];
  hardware: HardwarePinMap;
}

export type GetHistoryParams = {
  /**
   * @minimum 1
   * @maximum 500
   */
  limit?: number;
};

export type GetHistory200 = {
  readings: SensorReading[];
};

export type GetAlertsParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetAlerts200 = {
  alerts: AlertEvent[];
};