# lib/api-zod/src/generated/types/mqttConnectionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart RO purifier API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MqttConnectionStatus {
  configured: boolean;
  connected: boolean;
  brokerUrl?: string;
}