# lib/api-zod/src/generated/types/systemStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart RO purifier API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DatabaseConnectionStatus } from "./databaseConnectionStatus";
import type { HardwarePinMap } from "./hardwarePinMap";
import type { MqttConnectionStatus } from "./mqttConnectionStatus";
import type { SensorReading } from "./sensorReading";
import type { SystemStatusMode } from "./systemStatusMode";

export interface SystemStatus {
  mode: SystemStatusMode;
  threshold: number;
  pumpRule: string;
  latest: SensorReading;
  mqtt: MqttConnectionStatus;
  database: DatabaseConnectionStatus;
  topics: string[];
  hardware: HardwarePinMap;
}